[chore] 프로젝트 관련 파일 이동
