+++ src/main/java/org/sopt/makers/internal/internal/InternalApiService.java
@@ -11,9 +11,9 @@ import org.sopt.makers.internal.member.domain.enums.Part;
 import org.sopt.makers.internal.member.repository.MemberProfileQueryRepository;
 import org.sopt.makers.internal.member.repository.MemberRepository;
 import org.sopt.makers.internal.project.domain.Project;
-import org.sopt.makers.internal.project.dto.response.ProjectLinkDao;
-import org.sopt.makers.internal.project.dto.response.ProjectMemberDao;
-import org.sopt.makers.internal.project.dto.response.ProjectMemberVo;
+import org.sopt.makers.internal.project.dto.dao.ProjectLinkDao;
+import org.sopt.makers.internal.project.dto.dao.ProjectMemberDao;
+import org.sopt.makers.internal.project.dto.ProjectMemberVo;
 import org.sopt.makers.internal.exception.MemberHasNotProfileException;
 import org.sopt.makers.internal.exception.NotFoundDBEntityException;
 import org.sopt.makers.internal.member.mapper.MemberMapper;
