+++ src/main/java/org/sopt/makers/internal/project/controller/ProjectController.java
@@ -9,7 +9,9 @@ import lombok.val;
 import org.sopt.makers.internal.common.util.InfiniteScrollUtil;
 import org.sopt.makers.internal.internal.InternalMemberDetails;
 import org.sopt.makers.internal.project.domain.Project;
-import org.sopt.makers.internal.project.dto.response.*;
+import org.sopt.makers.internal.project.dto.response.allProject.ProjectAllResponse;
+import org.sopt.makers.internal.project.dto.response.detailProject.ProjectDetailResponse;
+import org.sopt.makers.internal.project.dto.response.allProject.ProjectResponse;
 import org.sopt.makers.internal.project.mapper.ProjectResponseMapper;
 import org.sopt.makers.internal.project.dto.request.ProjectSaveRequest;
 import org.sopt.makers.internal.project.dto.request.ProjectUpdateRequest;
