+++ src/main/java/org/sopt/makers/internal/project/dto/dao/ProjectLinkDao.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.internal.project.dto.response;
+package org.sopt.makers.internal.project.dto.dao;
 
 import com.querydsl.core.annotations.QueryProjection;
