+++ src/main/java/org/sopt/makers/internal/project/dto/response/allProject/ProjectMemberResponse.java
@@ -1,0 +1,7 @@
+package org.sopt.makers.internal.project.dto.response.allProject;
+
+public record ProjectMemberResponse(
+        Long memberId,
+        String memberName,
+        String memberProfileImage
+){}
