+++ src/main/java/org/sopt/makers/internal/project/dto/response/detailProject/ProjectLinkResponse.java
@@ -1,0 +1,7 @@
+package org.sopt.makers.internal.project.dto.response.detailProject;
+
+public record ProjectLinkResponse(
+        Long linkId,
+        String linkTitle,
+        String linkUrl
+){}
