+++ src/main/java/org/sopt/makers/internal/project/service/ProjectService.java
@@ -7,16 +7,16 @@ import org.sopt.makers.internal.exception.WrongImageInputException;
 import org.sopt.makers.internal.external.platform.InternalUserDetails;
 import org.sopt.makers.internal.external.platform.PlatformClient;
 import org.sopt.makers.internal.member.domain.Member;
-import org.sopt.makers.internal.member.domain.MemberSoptActivity;
 import org.sopt.makers.internal.member.repository.MemberRepository;
 import org.sopt.makers.internal.project.domain.ProjectLink;
 import org.sopt.makers.internal.project.domain.MemberProjectRelation;
 import org.sopt.makers.internal.project.domain.Project;
 import org.sopt.makers.internal.exception.ClientBadRequestException;
 import org.sopt.makers.internal.exception.NotFoundDBEntityException;
-import org.sopt.makers.internal.project.dto.response.*;
-import org.sopt.makers.internal.project.mapper.ProjectMapper;
-import org.sopt.makers.internal.member.repository.soptactivity.MemberSoptActivityRepository;
+import org.sopt.makers.internal.project.dto.response.allProject.ProjectMemberResponse;
+import org.sopt.makers.internal.project.dto.response.detailProject.ProjectDetailMemberResponse;
+import org.sopt.makers.internal.project.dto.response.detailProject.ProjectDetailResponse;
+import org.sopt.makers.internal.project.dto.response.allProject.ProjectResponse;
 import org.sopt.makers.internal.project.dto.request.ProjectSaveRequest;
 import org.sopt.makers.internal.project.dto.request.ProjectUpdateRequest;
 import org.sopt.makers.internal.project.mapper.ProjectResponseMapper;
